bug fixing. Get applied posts from users data with id

+++ resolvers/postResolver.js
@@ -8,7 +8,7 @@ export default {
   User: {
     applied_sports: async (parent, args) => {
       console.log(parent);
-      return Post.find({_id: {$in: parent.applied_sports}});
+      return Post.find({_id: {$in: parent.applied_sports}}).sort({_id: -1});
     },
   },
 
@@ -18,9 +18,14 @@ export default {
       if (!context.user) {
         throw new AuthenticationError('Not Authorized');
       }
-      return Post.find();
+      return Post.find().sort({_id: -1});
     },
-    post: async (parent, args) => await Post.findById(args.id),
+    post: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('Not Authorized');
+      }
+      return Post.findById(args.id);
+    },
     postByUser: async (parent, args, context) => {
       if (!context.user) {
         throw new AuthenticationError('Not Authorized');
@@ -40,15 +45,24 @@ export default {
       if (!context.user) {
         throw new AuthenticationError('Not Authorized');
       }
-      console.log('postInfo', args.postInfo);
+
       try {
         const newPost = await new Post(args.postInfo);
-        return newPost.save();
+        const updatedUser = await User.findOneAndUpdate(
+            {_id: args.postInfo.owner},
+            {$push: {applied_sports: newPost.id}},
+            {returnDocument: 'after'});
+        console.log('USER', args);
+        await updatedUser.save();
+        return await newPost.save();
       } catch (err) {
         throw new Error(err);
       }
     },
-    updatePost: async (parent, args) => {
+    updatePost: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('Not Authorized');
+      }
       try {
         const post = await Post.findOneAndUpdate(args.id, args.postInfo,
             {returnDocument: 'after'});
@@ -57,18 +71,23 @@ export default {
         throw new Error(err);
       }
     },
-    deletePost: async (parent, args) => {
+    deletePost: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('Not Authorized');
+      }
       try {
         return Post.findOneAndDelete(args.id);
       } catch (err) {
         throw new Error(err);
       }
     },
-    applyToPost: async (parent, args) => {
+    applyToPost: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('Not Authorized');
+      }
       try {
         const newParticipant = args.participantId;
-        console.log(newParticipant);
-        console.log('POSTID', args.id);
+
         const updatedPost = await Post.findOneAndUpdate({_id: args.id},
             {$push: {participants: newParticipant}},
             {returnDocument: 'after'});
@@ -77,19 +96,31 @@ export default {
             {_id: args.participantId},
             {$push: {applied_sports: args.id}},
             {returnDocument: 'after'});
+
         await updatedUser.save();
         return await updatedPost.save();
       } catch (err) {
         throw new Error(err);
       }
     },
-    leaveFromPost: async (parent, args) => {
+    leaveFromPost: async (parent, args, context) => {
+      if (!context.user) {
+        throw new AuthenticationError('Not Authorized');
+      }
       try {
         const participant = args.participantId;
-        console.log(participant);
-        const updatedPost = await Post.findOneAndUpdate(args.id,
+
+        const updatedPost = await Post.findOneAndUpdate(
+            {_id: args.id},
             {$pull: {participants: participant}},
             {returnDocument: 'after'});
+
+        const updatedUser = await User.findOneAndUpdate(
+            {_id: args.participantId},
+            {pull: {applied_sports: args.id}},
+            {returnDocument: 'after'});
+
+        await updatedUser.save();
         return await updatedPost.save();
       } catch
           (err) {
